requirements.txt
PyQt6>=6.5.0,<6.8.0 
Pillow>=9.0.0,<11.0.0
numpy>=1.20.0,<2.0.0
python-rtmidi>=1.4.0,<2.0.0
mss>=6.0.0,<10.0.0
appdirs>=1.4.0,<2.0.0